Address feedback about agent being able to switch between policies at runtime.
